// src/app/models/previousLeaveBalance.ts
export class PreviousLeaveBalance {
    id?: number;
    employeeNumber?: string;
    employeeNameFl?: string;
    employeeNameSl?: string;
    civilId?: string;
    religionId?: number;
    employeeId?: number;
    leaveId?: number;
    leaveNameFl?: string;
    leaveNameSl?: string;
    countryId?: number;
    previousBalance?: number;
    year?: number;
}